Bump Version 1.1.1000

+++ ttygif/tty/image.pxd
@@ -45,7 +45,7 @@ cdef class image:
     cdef int         length
     cdef rect get_rect(self)
     cdef array.array create_buffer(self,size,init_value)
-    cdef uint8_t get_position(self,int x,int y)
+    cdef uint32_t get_position(self,int x,int y)
     cdef get_pixel(self,int x,int y)
     cdef uint8_t get_pixel_1byte(self,int x,int y)
     cdef void get_pixel_3byte(self,int x,int y,uint8_t[3] element)
